+++ js/editor.js
@@ -107,6 +107,11 @@
             url: configUrl,
             success: function onSuccess(config) {
                 if (config) {
+                    if (config.redirectUrl) {
+                        location.href = config.redirectUrl;
+                        return;
+                    }
+
                     if (config.error != null) {
                         OCA.Onlyoffice.showMessage(config.error, "error", {timeout: -1});
                         return;
@@ -115,11 +120,6 @@
                     if ((config.document.fileType === "docxf" || config.document.fileType === "oform")
                         && docsVersion[0] < 7) {
                         OCA.Onlyoffice.showMessage(t(OCA.Onlyoffice.AppName, "Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online"), "error", {timeout: -1});
-                        return;
-                    }
-
-                    if (config.redirectUrl) {
-                        location.href = config.redirectUrl;
                         return;
                     }