+++ templates/settings.php
@@ -126,6 +126,12 @@
         <button id="onlyofficeClearVersionHistory" class="button"><?php p($l->t("Clear")) ?></button>
     </p>
 
+    <p>
+        <input type="checkbox" class="checkbox" id="onlyofficePlugins"
+            <?php if ($_["plugins"]) { ?>checked="checked"<?php } ?> />
+        <label for="onlyofficePlugins"><?php p($l->t("Enable plugins")) ?></label>
+    </p>
+
     <p class="onlyoffice-header"><?php p($l->t("The default application for opening the format")) ?></p>
     <div class="onlyoffice-exts">
         <?php foreach ($_["formats"] as $format => $setting) { ?>