+++ controller/editorapicontroller.php
@@ -377,7 +377,8 @@ class EditorApiController extends OCSController {
             if (method_exists(IShare::class, "getAttributes")) {
                 $share = empty($share) ? $fileStorage->getShare() : $share;
                 $attributes = $share->getAttributes();
-                if ($attributes !== null && !$attributes->getAttribute("permissions", "download")) {
+                $downloadAttr = isset($attributes) ? $attributes->getAttribute("permissions", "download") : null;
+                if (isset($downloadAttr) && !$downloadAttr) {
                     $canDownload = false;
                 }
             }