+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## Added
 - disable plugins setting
 
+##
 ## Changed
 - fix download permission
 
+##
 ## Added
 - Dutch translation
 - Chinese (Traditional, Taiwan), Basque (Spain) empty file templates